538-djangodg___more/538-c-django-dg-folder-crispy-partiallyworking/538b-django-docker/djangocode/requirements.txt
Django>=2,<3
django-extensions
djangorestframework
django-crispy-forms
